requirements.txt
# Python 3.10 / SDRE / Compute Canada friendly baseline requirements.
# OpenCV is intentionally omitted. Load OpenCV as a site module if needed.
# On clusters, install the CUDA-matched torch/torchvision build your site supports.
torch>=2.1,<2.6
torchvision>=0.16,<0.21
timm>=1.0.18,<1.1
huggingface_hub>=0.24,<1.0
safetensors>=0.4,<0.7
numpy>=1.24,<2.3
pandas>=2.0,<3.0
matplotlib>=3.7,<3.11
scikit-learn>=1.3,<1.8
scipy>=1.11,<1.16
pillow>=10.0,<13
tqdm>=4.66,<5
pyyaml>=6.0,<7
openpyxl>=3.1,<3.2
xlsxwriter>=3.1,<4
shap>=0.45,<0.50

# Optional advanced-analysis extras. The code falls back safely when these are absent.
# pacmap>=0.7,<0.9
# trimap>=1.1,<2
# plotly>=5.20,<6
